packages/tgsdk/tgsdk-1.3.0.tar.gz/tgsdk-1.3.0/tgsdk/entities/payments/answerprecheckoutquery.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

# Copyright (c) 2015-2022 Evgeniy Privalov, https://linkedin.com/in/evgeniyprivalov/

from typing import Any

from tgsdk import TelegramEntity


class AnswerPreCheckoutQuery(TelegramEntity):
	"""
	https://core.telegram.org/bots/api/#answerprecheckoutquery

	"""

	__slots__ = ("pre_checkout_query_id", "ok", "error_message")

	def __init__(
		self,
		pre_checkout_query_id: str,
		ok: bool,
		error_message: str,

		**_kwargs: Any
	):
		self.pre_checkout_query_id = pre_checkout_query_id
		self.ok = ok
		self.error_message = error_message
